+++ lib/core/states/app_settings_state.dart
@@ -3,7 +3,7 @@ import 'package:signals_flutter/signals_flutter.dart';
 /// 应用设置状态
 class AppSettingsState {
   // 应用名称
-  final appName = signal('Astral-ng');
+  final appName = signal('AstralNG');
 
   // 启用连接状态栏通知
   final enableConnectionNotification = signal(true);